requirements.txt
APScheduler~=3.11
PyYAML~=6.0
adafruit-circuitpython-ads1x15~=3.0
RPi.GPIO~=0.7
luma.oled~=3.14